Move functions to docker-common
